refactor: move logging extensions and rename access logging
